refactor: AsyncAiService의 메서드를 AiSerivce로 통합 (#499)
