--- backend/src/main/java/com/tamnara/backend/news/service/AsyncAIService.java
@@ -1,11 +1,0 @@
-package com.tamnara.backend.news.service;
-
-import com.tamnara.backend.news.dto.StatisticsDTO;
-import com.tamnara.backend.global.dto.WrappedDTO;
-
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-
-public interface AsyncAIService {
-    CompletableFuture<WrappedDTO<StatisticsDTO>> getAIStatistics(List<String> keywords);
-}
